Extract SnapshotVolumeAction from the actioner
